+++ src/main/java/org/traccar/broadcast/BaseBroadcastService.java
@@ -21,6 +21,7 @@ import java.util.Set;
 import org.traccar.model.BaseModel;
 import org.traccar.model.Device;
 import org.traccar.model.Event;
+import org.traccar.model.ObjectOperation;
 import org.traccar.model.Permission;
 import org.traccar.model.Position;
