+++ src/main/java/org/traccar/session/cache/CacheManager.java
@@ -19,7 +19,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.traccar.broadcast.BroadcastInterface;
 import org.traccar.broadcast.BroadcastService;
-import org.traccar.broadcast.ObjectOperation;
+import org.traccar.model.ObjectOperation;
 import org.traccar.config.Config;
 import org.traccar.model.Attribute;
 import org.traccar.model.BaseModel;
